fix: change vuex generation to use a default TS client path (#2831)
* fix: change TS client generation to use a default config generating vuex

With the latest changes it is possible to only define a config path for
the Vuex generation leaving the TS client path empty. This changeset
makes sure that in that case there is a default value for it.

* chore: update changelog

* chore: check that package.json exists before trying to update it

* chore: clean the configured TS client path

* refactor: add output path support ts-client generation cmd

* refactor: change `GenerateTSClient` to accept the output path

This change keeps the semantics of the `Generate` function targets and
allows overriding the configured or default output path when generating
the typescript client.

* fix: change vuex dependency update to handle non absolute paths

The TS client folder can be a name in which case it should be considered
a relative path, which is the case for the default directory name
"ts-client".