test: integration build the ignite binary (#2639)
* test: integration build the ignite binary (WIP)

Previously, the ignite binary was supposed to be present in the path,
which can be error prone if the binary is not up to date with the
integration version.

The change builds the binary before the integration are run, using a
init() function. This is not very adequate since the binary is not
removed at the end of the test. To be able to remove it at the end of
the tests we could:

1) add a TestMain in each integration sub-package: not super fan of this
   because it involves to repeat the same code for each package
2) move all tests in the same package: preferable but delete the
   separation between the tests.

* test: constant ignite binary location

Also:
- use sync.Once rather than init() func
- use gocmd rather than exec.Command directly

* chore: no longer need to build the binary in gh action

* change cl section

Co-authored-by: Alex Johnson <alex@shmeeload.xyz>