fix: test and test script (#3271)
* fix: test and test script

One test was actually failing on main but this wasn't detected by the CI
because the script `script/test-coverage` returns 0 even if the tests
fail.

This change fix the test and the script.

* chore: change CLI to display errors in red next to an error icon

* chore: add missing red color styling to default CLI error messages

* rename

* refactor to remove cliui

* fix test script

When no match, grep returns 1, when xargs command return an error, xargs
return 123. Because of `set -e`, all commands must return 0, which is
good IMO. The alternative is to just check $? after the `go test`, but
then some errors would be ignored in the remaining of the script.

Also fix the test fix on `config/plugins.ParseDir`. There's no need to
have an empty file in the empty dir.

* try to find what part of the script has an error

* fix test, empty dir was somehow skipped from the CI

* ensure plugin load have access to output outside of plugin commands

Co-authored-by: jeronimoalbi <jeronimo.albi@tendermint.com>
Co-authored-by: Alex Johnson <alex@shmeeload.xyz>